Притяжение убывает обратно пропорционально квадрату расстояния

+++ Ssz.AI/Views/Model05View.axaml.cs
@@ -204,7 +204,7 @@ public partial class Model05View : UserControl
     private void Refresh_ImagesSet1()
     {
         double position = PositionScrollBar.Value;
-        double angle = MathHelper.DegreesToRadians(AngleScrollBar.Value);
+        double angle = MathHelper.DegreesToRadians((float)AngleScrollBar.Value);
         ImagesSet1.MainItemsControl.ItemsSource = Model.GetImageWithDescs1(position, angle);
 
         PositionTextBlock.Text =
